pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = ""
authors = ["okhaled9 <okhaled9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
pytz = "^2023.3"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
black = "^23.9.1"
pillow = "^10.0.1"
django-debug-toolbar = "^4.2.0"
django-imagekit = "^5.0.0"
django-cleanup = "^8.0.0"
django-rest-knox = "^4.2.0"
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
